Store full 64-byte shielded address in delegations map
Replaces the truncated Field value with a ShieldedAddressData struct
wrapping Bytes<64>, so the real mn_shield-addr bech32m (coin_pub_key ||
enc_pub_key) can round-trip between the frontend wallet linker, the
contract, and the backend state machine without any byte loss or
format-mismatched display.

Contract:
- ShieldedAddressData { data: Bytes<64> } struct
- delegations: Map<Field, ShieldedAddressData>
- register_delegation(wallet_address: Bytes<64>)

Frontend:
- registerDelegation takes a 64-byte Uint8Array
- myDelegatedAddress derived as Uint8Array | null
- Wallet link decodes addresses.shieldedAddress bech32m to raw bytes
- Boot encodes the 64 bytes straight to mn_shield-addr for display
- New toBech32mShieldAddr helper in bech32-utils

Backend state machine:
- bytesToBech32Shield encoder
- syncDelegations parses the indexer's decimal BigInt representation of
  the Bytes<64> value, converts to BE hex, reverses to natural byte
  order, and bech32m-encodes with the shield-addr prefix

Requires contract redeployment and TRUNCATE public.d2d_delegations
since the ledger storage type has changed.

+++ frontend/src/api/src/common-types.ts
@@ -75,6 +75,6 @@ export type Game2DerivedState = {
   playerBossProgress: Map<bigint, Map<bigint, boolean>>;
   // quest durations in seconds per level: biome -> difficulty -> seconds (missing means default 1200)
   questDurations: Map<bigint, Map<bigint, bigint>>;
-  /** Current player's delegated wallet address, or null if not delegated. */
-  myDelegatedAddress: bigint | null;
+  /** Current player's delegated wallet shielded address (64 raw bytes: coin_pk || enc_pk), or null if not delegated. */
+  myDelegatedAddress: Uint8Array | null;
 }
